Refactoring __init__.py to avoid cyclic dependency
File __init__.py contains the widely used functions MODULE_LIST and
SCRIPT_LIST. However, since this file is imported in all over the
project, it causes cyclic dependency very often. E.g. in order to import
lib/templates.py in lib/compile.json, a cycle is encountered -
lib/compile.py -> lib/templates.py -> __init__.py -> lib/compile.py.
These two functions have been refactored into a new script located at
lib/scripts.py.

The constant components of __init__.py such as HOME_DIR etc. are also
referenced much often. These variables have also been refactored into
lib/defaults.py.

The references to these variables and functions have been modified
throughout the project to keep it consistent. THe documentation has been
modified as well to accomodate the newly generated lib/scripts.py
module.
