+++ retriever/engines/msaccess.py
@@ -2,7 +2,8 @@ from __future__ import print_function
 from builtins import str
 import os
 from retriever.lib.models import Engine, no_cleanup
-from retriever import DATA_DIR, current_platform
+from retriever import current_platform
+from retriever.lib.defaults import DATA_DIR
 
 
 class engine(Engine):
