+++ retriever/lib/templates.py
@@ -4,9 +4,9 @@ from __future__ import print_function
 from builtins import object
 import os
 import shutil
-from retriever import DATA_DIR
+from retriever.lib.defaults import DATA_DIR
 from retriever.lib.models import *
-from retriever.lib.tools import choose_engine
+from retriever.engines import choose_engine
 
 
 class Script(object):
